% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/featureSelect.R
\name{featureSelect}
\alias{featureSelect}
\title{Feature Selection}
\usage{
featureSelect(input, batch_effect = NULL, VGs = NULL)
}
\arguments{
\item{input}{\code{SpatialExperiment} or \code{SingleCellExperiment}: Input
data which can be either \code{SpatialExperiment} or
\code{SingleCellExperiment}. It is assumed to have an \code{assays} slot
containing \code{counts} assay for \code{devianceFeatureSelection()} to
successfully operate and calculate the deviance and rank. The
\code{logcounts} is strongly recommended to be included in \code{assays}.
Also, the input data is assumed to set up \code{rownames(input)} as
\code{genes} and \code{gene_name} in \code{rowData(input)}.}

\item{batch_effect}{\code{character}: Either batch on \code{slide} or
\code{subject} based on what metadata is available within input data.
The name of \code{slide} or \code{subject} within each input object
should be specified based on different scenarios.}

\item{VGs}{\code{character}: Highly Variable Genes (HVGs) for
\code{SingleCelleExperiment} object or Spatially Variable Genes (SVGs)
for \code{SpatialExperiment} object. If it is a data frame, it is assumed
to contain one column of identified variable genes with column name as
"gene_name".}
}
\value{
If the input was provided as a \code{SpatialExperiment} or
\code{SingleCellExperiment} object, the output values are returned as
a data frame containing the deviance and rank with and without the bacth
effect, the corresponding difference, the corresponding nSD, gene, gene
name, and whether the gene is outlier defined by the chosen deviance and
rank cutoff.
}
\description{
Function to conduct feature selection on snRNA-seq and spatial
transcriptomics data and calculate the difference of deviance and rank
with and without the selected batch effect.
}
\examples{
# library(spatialLIBD)
# spe <- fetch_data(type = "spe")

# featureSelect(spe, batch_effect = "brain", VGs = SVGs)

}
