% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spe_svgs.R
\docType{data}
\name{spe}
\alias{spe}
\title{Human Hippocampus .}
\format{
An SpatialExperiment object.
\describe{}
}
\source{
\href{https://github.com/LieberInstitute/spatial_hpc}{spatialHPC}
}
\usage{
data(spe_svgs)
}
\description{
The original data was generated spatially-resolved transcriptomics (SRT) and
single-nucleus RNA-sequencing (snRNA-seq) data from adjacent tissue sections
of the anterior human hippocampus across ten adult neurotypical donors.
SRT data was generated using 10x Genomics Visium (n=36 capture areas) and
10x Genomics Visium Spatial Proteogenomics (SPG) (n=8 capture areas).
snRNA-seq data was generated using 10x Genomics Chromium (n=26 total
snRNA-seq libraries). The original data was processed using \code{nnSVG()} feature
selection, and \code{spe_svgs} dataset is the \code{SpatialExperiment} object
containing the 2098 spatially variable genes (SVGs).
}
\examples{
data(spe_svgs)
}
\keyword{datasets}
