% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biasDetect.R
\name{biasDetect}
\alias{biasDetect}
\title{biasDetect}
\usage{
biasDetect(input, batch_effect, nSD_dev = 5, nSD_rank = 5, visual = FALSE)
}
\arguments{
\item{input}{Input data: assumed to be formatted as a
\code{SpatialExperiment} object or \code{SingleCellExperiment} with an
\code{assays} slot named \code{counts} containing raw expression counts.}

\item{batch_effect}{Batch effect: either \code{slide} or \code{subject} based
on what metadata is available within input data. The name of \code{slide}
or \code{subject} within each input object can be different.}

\item{nSD_dev}{number of standard deviation (nSD) on deviance difference}

\item{nSD_rank}{number of standard deviation (nSD) on rank difference}

\item{visual}{visualization of nSD on deviance or rank difference. Default F}
}
\value{
If the input was provided as a \code{SpatialExperiment} object, the
output values are returned as additional columns in the \code{rowData} slot
of the input object. If the input was provided as a \code{numeric} matrix
of values, the output is returned as a \code{numeric} matrix. The output
values include spatial variance parameter estimates, likelihood ratio (LR)
statistics, effect sizes (proportion of spatial variance), p-values, and
multiple testing adjusted p-values.
}
\description{
Function to run 'BiasDetect' method for spatially variable genes (SVGs)
identification on spatial-resolved transcriptomics (SRT) data with some types
of batch effect.
}
\details{
\code{BiasDetect} is a feature-based Quality Control (QC) to identify SVGs on
spatial transcriptomics data with sometypes of batch effect. Regarding to the
spatial transcriptomics data experients, the batch can be defined as
"slide" or "subject".The \code{BiasDetect} method is based on binomial deviance
model (Townes et al, 2019) and applies cutoffs based on the number of
standard deviation (nSD) of deviance and rank difference as the data-driven
thresholding approach to detect the batch-biased features.
}
