% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biasDetect.R
\name{biasDetect}
\alias{biasDetect}
\title{Find Bias Genes}
\usage{
biasDetect(batch_df, nSD_dev = 5, nSD_rank = 5, visual = FALSE)
}
\arguments{
\item{batch_df}{\code{data.frame} : Input data frame generated from
\code{featureSelection()} function using \code{SpatialExperiment} or
\code{SingleCellExperiment} object containing the raw data.}

\item{nSD_dev}{\code{integer}: Number of standard deviation (nSD) on deviance
difference. The default value is 5.}

\item{nSD_rank}{\code{integer}: Number of standard deviation (nSD) on rank
difference. The default value is 5.}

\item{visual}{\code{logical}: Whether to display the detected bias genes by
visualizations of \code{deviance} and \code{rank}. Default = FALSE to
return the data frame. If it is TRUE, the returned format will be
two parallel plots presenting bias genes based on both \code{nSD_dev} and
\code{nSD_rank}.}
}
\value{
The output values will be the identified bias genes. The returned
format can be either a data frame or a list with two elements, giving
option to see either \code{deviance} or \code{rank} plot.
}
\description{
Function to filter out the bias genes based on selected
selected threshold of nSD of deviance and rank in data frame or plot.
}
\examples{
data(spe)
spe_svgs <- spe

# biasDetect()

}
