% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biasDetect.R
\name{featureSelect}
\alias{featureSelect}
\alias{biasDetect}
\title{Feature Selection}
\usage{
featureSelect(input, batch_effect = NULL, VGs = NULL)

biasDetect(batch_df, nSD_dev = 5, nSD_rank = 5, visual = FALSE)
}
\arguments{
\item{input}{\code{SpatialExperiment} or \code{SingleCellExperiment}: Input
data which can be either \code{SpatialExperiment} or
\code{SingleCellExperiment}. It is assumed to have an \code{assays} slot
containing \code{counts} assay for \code{devianceFeatureSelection()} to
successfully operate and calculate the deviance and rank. The
\code{logcounts} is strongly recommended to be included in \code{assays}.
Also, the input data is assumed to set up \code{rownames(input)} as
\code{genes} and \code{gene_name} in \code{rowData(input)}.}

\item{batch_effect}{\code{character}: Either batch on \code{slide} or
\code{subject} based on what metadata is available within input data.
The name of \code{slide} or \code{subject} within each input object
should be specified based on different scenarios.}

\item{VGs}{\code{character}: Highly Variable Genes (HVGs) for
\code{SingleCelleExperiment} object or Spatially Variable Genes (SVGs)
for \code{SpatialExperiment} object. If it is a data frame, it is assumed
to contain one column of identified variable genes with column name as
"gene_name".}

\item{batch_df}{\code{data.frame} : Input data frame generated from
\code{featureSelection()} function using \code{SpatialExperiment} or
\code{SingleCellExperiment} object containing the raw data.}

\item{nSD_dev}{\code{integer}: Number of standard deviation (nSD) on deviance
difference. The default value is 5.}

\item{nSD_rank}{\code{integer}: Number of standard deviation (nSD) on rank
difference. The default value is 5.}

\item{visual}{\code{logical}: Whether to display the detected bias genes by
visualizations of \code{deviance} and \code{rank}. Default = FALSE to
return the data frame. If it is TRUE, the returned format will be
two parallel plots presenting bias genes based on both \code{nSD_dev} and
\code{nSD_rank}.}
}
\value{
If the input was provided as a \code{SpatialExperiment} or
\code{SingleCellExperiment} object, the output values are returned as
a data frame containing the deviance and rank with and without the bacth
effect, the corresponding difference, the corresponding nSD, gene, gene
name, and whether the gene is outlier defined by the chosen deviance and
rank cutoff.

The output values will be the identified bias genes. The returned
format can be either a data frame or a list with two elements, giving
option to see either \code{deviance} or \code{rank} plot.
}
\description{
Function to conduct feature selection on snRNA-seq and spatial
transcriptomics data and calculate the difference of deviance and rank
with and without the selected batch effect.

Function to filter out the bias genes based on selected
selected threshold of nSD of deviance and rank in data frame or plot.
}
\examples{
# library(spatialLIBD)
# spe <- fetch_data(type = "spe")

# featureSelect(spe, batch_effect = "brain", VGs = SVGs)

# library(spatialLIBD)

# batch_df <- featureSelect(spe, batch_effect = "brain", VGs = SVGs)
# biasDetect(batch_df, nSD_dev = 5, nSD_rank = 6)
# biasDetect(batch_df, nSD_dev = 5, nSD_rank = 6, visual = TRUE)

}
